backend/requirements.txt
Django>=6.0,<7.0
djangorestframework==3.15.2
django-cors-headers==4.6.0
Pillow>=11.0.0
python-dotenv==1.0.1
django-ckeditor==6.7.1
django-filter==24.3
django-unfold==0.50.0
django-modeltranslation==0.18.11
